+++ frameworks/core/components_ng/pattern/text_field/text_field_pattern.cpp
@@ -3245,7 +3245,7 @@ void TextFieldPattern::InsertValue(const std::string& insertValue)
     if (inputValue == DEFAULT_MODE) {
         originLength = originLength + ONE_CHARACTER;
     }
-    if (textFieldLayoutProperty->GetShowCounterValue(false) && originLength == maxlength &&
+    if (textFieldLayoutProperty->GetShowCounterValue(false) && originLength == static_cast<int32_t>(maxlength) &&
         inputValue == DEFAULT_MODE) {
         UpdateCounterBorderStyle(originLength, maxlength);
     }
@@ -3328,7 +3328,7 @@ void TextFieldPattern::UpdateCounterBorderStyle(int32_t& textLength, uint32_t& m
     auto showBorder = textFieldLayoutProperty->GetShowHighlightBorderValue(true);
     if (static_cast<uint32_t>(textLength) >= (maxLength - ONE_CHARACTER) && !IsTextArea() && showBorder == true) {
         SetUnderlineColor(theme->GetErrorUnderlineColor());
-    } else if (textLength >= maxLength && IsTextArea() && showBorder == true) {
+    } else if (textLength >= static_cast<int32_t>(maxLength) && IsTextArea() && showBorder == true) {
         HandleCounterBorder();
     }
     return;
